refactor: DTO 명명 구체화
